+++ product-service/src/controllers/getByID.controller.ts
@@ -1,0 +1,13 @@
+import { APIGatewayProxyHandler } from "aws-lambda";
+
+import getByID from "../services/getByID.service";
+import { generateErrorResponse, generateSuccessResponse } from "../helpers";
+
+export const getProductById: APIGatewayProxyHandler = async (event) => {
+  try {
+    const result = getByID(event.pathParameters.productId);
+    return generateSuccessResponse(result);
+  } catch (e) {
+    return generateErrorResponse(e);
+  }
+};
