--- product-service/src/requestHandlers/getAll.ts
@@ -1,12 +1,0 @@
-import { APIGatewayProxyHandler } from "aws-lambda";
-
-import products from "../data/products";
-import { getHeaders } from "../helpers/index";
-
-export const getAllProducts: APIGatewayProxyHandler = async () => {
-  return {
-    statusCode: 200,
-    body: JSON.stringify(products),
-    headers: getHeaders(),
-  };
-};
