+++ product-service/src/services/getAll.service.ts
@@ -1,0 +1,6 @@
+import products from "../data/products";
+
+export default async () => {
+    return [...products]
+  };
+
