+++ product-service/src/services/getByID.service.ts
@@ -1,0 +1,10 @@
+import products from "../data/products";
+
+export default async (id: string) => {
+    const product = products.find((a) => a.id === id);
+    if (!product) {
+      throw new Error("Product not found");
+    }
+    return product
+  };
+
